11.1HD/requirements.txt
Flask
paho-mqtt
geopy
gpiozero
